src/requirements.txt
Flask==3.1.3
Werkzeug==3.1.8
python-magic==0.4.27
python-magic-bin==0.4.14;sys_platform == "win32"
gunicorn==25.3.0; sys_platform != "win32"